test: split CLI tests into command-focused files
